Move helper files for CommunityGAN driver into root dir
