+++ applications/graph/communityGAN/default.config
@@ -1,7 +1,9 @@
-[Defaults]
 [Graph]
-graph_file = /usr/workspace/llamag/lbann/applications/graph/communityGAN/prunejuice_data/input_edge_list/four_clique_x2_edge_list
-graph_store = /dev/shm/gr/
+# Edge-list file
+file = /usr/workspace/llamag/lbann/applications/graph/communityGAN/prunejuice_data/input_edge_list/four_clique_x2_edge_list
+num_vertices = 0
+# Where to store ingested graph
+ingest_dir = /dev/shm/gr/
 
 [Pattern]
 pattern_in_dir = /usr/workspace/llamag/lbann/applications/graph/communityGAN/prunejuice_data/unlabeled_patterns_0001/4_clique_unlabeled_apm/0_0/pattern/
@@ -15,10 +17,11 @@ create_motif_file = False
 motif_set_file = Motif_set
 
 [RW]
+#out_dir = #
+
 rw_walks_store = /usr/workspace/llamag/lbann/applications/graph/communityGAN/largescale_node2vec/.gr
 rw_out_filename = /usr/workspace/llamag/lbann/applications/graph/communityGAN/driver/RW-out
 rw_walk_len = 80
 rw_num_walkers = 40
 rw_p = 1.0
 rw_q = 1.0
-
